requirements.txt
google-genai>=1.0.0
PyGithub>=2.1.1
gitpython>=3.1.40
rich>=13.7.0
schedule>=1.2.1
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.3
pytest>=8.0.0
